Re-base gascheck (#252)
Change the base for "gascheck"
This is not a change in the contracts, but in the scripts to calculate the gas usage.
The default sample call now calls entryPoint() in the account, which returns a constant (entrypoint) instead of a variable (nonce)
The 2000 extra gas overhead seen results/gas-checker.txt in gas is actually a fix of a mistake: previously, accessing the "nonce" field was attributed to the "exec" call (because we called "nonce()"). Now it is rightfully attributed to the validation, not execution.